+++ src/index.ts
@@ -1,4 +1,5 @@
 export * from './client';
 export * from './mvm';
+export * from './webview';
 export * from './mixin/nfo';
 export * from './mixin/dump_transacion';
